+++ R/03-modelResults2DKernel.R
@@ -17,6 +17,7 @@ modelResults2DKernelUI <- function(id, title = "", asFruitsTab = FALSE){
       # left sidebar ----
       sidebarPanel(
         width = 2,
+        style = "position:fixed; width:14%; max-width:220px; overflow-y:auto; height:88%",
         selectInput(ns("dataSource"),
                     "Data source",
                     choices = if (!asFruitsTab) c("Database" = "db", "Upload file" = "file", "Saved map" = "model") else c("Database" = "db"),
@@ -180,6 +181,7 @@ modelResults2DKernelUI <- function(id, title = "", asFruitsTab = FALSE){
       # right sidebar ----
         sidebarPanel(
           width = 2,
+          style = "position:fixed; width:14%; max-width:220px; overflow-y:auto; height:88%",
           radioButtons(inputId = ns("Centering"),
                        label = "Map Centering",
                        choices = c("0th meridian" = "Europe", "160th meridian" = "Pacific")),